+++ bindings/nodejs/src/lib.rs
@@ -12,66 +12,71 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#![deny(clippy::all)]
-
 #[macro_use]
 extern crate napi_derive;
 
 use std::collections::HashMap;
-use std::str;
+use std::str::FromStr;
 
 use futures::TryStreamExt;
 use napi::bindgen_prelude::*;
-use opendal::Builder;
 use time::format_description::well_known::Rfc3339;
 
-#[napi]
-#[derive(Debug, Eq, PartialEq)]
-pub enum Scheme {
-    /// [azblob][crate::services::Azblob]: Azure Storage Blob services.
-    Azblob,
-    /// [azdfs][crate::services::Azdfs]: Azure Data Lake Storage Gen2.
-    Azdfs,
-    /// [dashmap][crate::services::Dashmap]: dashmap backend support.
-    Dashmap,
-    /// [fs][crate::services::Fs]: POSIX alike file system.
-    Fs,
-    /// [gcs][crate::services::Gcs]: Google Cloud Storage backend.
-    Gcs,
-    /// [ghac][crate::services::Ghac]: Github Action Cache services.
-    Ghac,
-    /// [hdfs][crate::services::Hdfs]: Hadoop Distributed File System.
-    Hdfs,
-    /// [http][crate::services::Http]: HTTP backend.
-    Http,
-    /// [ftp][crate::services::Ftp]: FTP backend.
-    Ftp,
-    /// [ipmfs][crate::services::Ipfs]: IPFS HTTP Gateway
-    Ipfs,
-    /// [ipmfs][crate::services::Ipmfs]: IPFS mutable file system
-    Ipmfs,
-    /// [memcached][crate::services::Memcached]: Memcached service support.
-    Memcached,
-    /// [memory][crate::services::Memory]: In memory backend support.
-    Memory,
-    /// [moka][crate::services::Moka]: moka backend support.
-    Moka,
-    /// [obs][crate::services::Obs]: Huawei Cloud OBS services.
-    Obs,
-    /// [oss][crate::services::Oss]: Aliyun Object Storage Services
-    Oss,
-    /// [redis][crate::services::Redis]: Redis services
-    Redis,
-    /// [rocksdb][crate::services::Rocksdb]: RocksDB services
-    Rocksdb,
-    /// [s3][crate::services::S3]: AWS S3 alike services.
-    S3,
-    /// [sled][crate::services::Sled]: Sled services
-    Sled,
-    /// [webdav][crate::services::Webdav]: WebDAV support.
-    Webdav,
-    /// [webhdfs][crate::services::Webhdfs]: WebHDFS RESTful API Services
-    Webhdfs,
+fn build_operator(
+    scheme: opendal::Scheme,
+    map: HashMap<String, String>,
+) -> Result<opendal::Operator> {
+    use opendal::services::*;
+
+    let op = match scheme {
+        opendal::Scheme::Azblob => opendal::Operator::from_map::<Azblob>(map)
+            .map_err(format_napi_error)?
+            .finish(),
+        opendal::Scheme::Azdfs => opendal::Operator::from_map::<Azdfs>(map)
+            .map_err(format_napi_error)?
+            .finish(),
+        opendal::Scheme::Fs => opendal::Operator::from_map::<Fs>(map)
+            .map_err(format_napi_error)?
+            .finish(),
+        opendal::Scheme::Gcs => opendal::Operator::from_map::<Gcs>(map)
+            .map_err(format_napi_error)?
+            .finish(),
+        opendal::Scheme::Ghac => opendal::Operator::from_map::<Ghac>(map)
+            .map_err(format_napi_error)?
+            .finish(),
+        opendal::Scheme::Http => opendal::Operator::from_map::<Http>(map)
+            .map_err(format_napi_error)?
+            .finish(),
+        opendal::Scheme::Ipmfs => opendal::Operator::from_map::<Ipmfs>(map)
+            .map_err(format_napi_error)?
+            .finish(),
+        opendal::Scheme::Memory => opendal::Operator::from_map::<Memory>(map)
+            .map_err(format_napi_error)?
+            .finish(),
+        opendal::Scheme::Obs => opendal::Operator::from_map::<Obs>(map)
+            .map_err(format_napi_error)?
+            .finish(),
+        opendal::Scheme::Oss => opendal::Operator::from_map::<Oss>(map)
+            .map_err(format_napi_error)?
+            .finish(),
+        opendal::Scheme::S3 => opendal::Operator::from_map::<S3>(map)
+            .map_err(format_napi_error)?
+            .finish(),
+        opendal::Scheme::Webdav => opendal::Operator::from_map::<Webdav>(map)
+            .map_err(format_napi_error)?
+            .finish(),
+        opendal::Scheme::Webhdfs => opendal::Operator::from_map::<Webhdfs>(map)
+            .map_err(format_napi_error)?
+            .finish(),
+        _ => {
+            return Err(format_napi_error(opendal::Error::new(
+                opendal::ErrorKind::Unexpected,
+                "not supported scheme",
+            )))
+        }
+    };
+
+    Ok(op)
 }
 
 #[napi]
@@ -80,66 +85,16 @@ pub struct Operator(opendal::Operator);
 #[napi]
 impl Operator {
     #[napi(constructor)]
-    pub fn new(service_type: Scheme, options: Option<HashMap<String, String>>) -> Result<Self> {
-        let ops = options.unwrap_or_default();
-        match service_type {
-            Scheme::Azblob => Ok(Self(
-                opendal::Operator::new(opendal::services::Azblob::from_map(ops))
-                    .map_err(format_napi_error)?
-                    .finish(),
-            )),
-            Scheme::Azdfs => Ok(Self(
-                opendal::Operator::new(opendal::services::Azdfs::from_map(ops))
-                    .map_err(format_napi_error)?
-                    .finish(),
-            )),
-            Scheme::Gcs => Ok(Self(
-                opendal::Operator::new(opendal::services::Gcs::from_map(ops))
-                    .map_err(format_napi_error)?
-                    .finish(),
-            )),
-            Scheme::Fs => Ok(Self(
-                opendal::Operator::new(opendal::services::Fs::from_map(ops))
-                    .map_err(format_napi_error)?
-                    .finish(),
-            )),
-            Scheme::Ghac => Ok(Self(
-                opendal::Operator::new(opendal::services::Ghac::from_map(ops))
-                    .map_err(format_napi_error)?
-                    .finish(),
-            )),
-            Scheme::Http => Ok(Self(
-                opendal::Operator::new(opendal::services::Http::from_map(ops))
-                    .map_err(format_napi_error)?
-                    .finish(),
-            )),
-            Scheme::Memory => Ok(Self(
-                opendal::Operator::new(opendal::services::Memory::default())
-                    .map_err(format_napi_error)?
-                    .finish(),
-            )),
-            Scheme::Obs => Ok(Self(
-                opendal::Operator::new(opendal::services::Obs::from_map(ops))
-                    .map_err(format_napi_error)?
-                    .finish(),
-            )),
-            Scheme::S3 => Ok(Self(
-                opendal::Operator::new(opendal::services::S3::from_map(ops))
-                    .map_err(format_napi_error)?
-                    .finish(),
-            )),
-            Scheme::Webdav => Ok(Self(
-                opendal::Operator::new(opendal::services::Webdav::from_map(ops))
-                    .map_err(format_napi_error)?
-                    .finish(),
-            )),
-            Scheme::Webhdfs => Ok(Self(
-                opendal::Operator::new(opendal::services::Webhdfs::from_map(ops))
-                    .map_err(format_napi_error)?
-                    .finish(),
-            )),
-            _ => Err(Error::from_reason("wrong operator type")),
-        }
+    pub fn new(scheme: String, options: Option<HashMap<String, String>>) -> Result<Self> {
+        let scheme = opendal::Scheme::from_str(&scheme)
+            .map_err(|err| {
+                opendal::Error::new(opendal::ErrorKind::Unexpected, "not supported scheme")
+                    .set_source(err)
+            })
+            .map_err(format_napi_error)?;
+        let options = options.unwrap_or_default();
+
+        build_operator(scheme, options).map(Operator)
     }
 
     #[napi]
@@ -154,6 +109,19 @@ impl Operator {
         let meta = self.0.blocking().stat(&path).map_err(format_napi_error)?;
 
         Ok(Metadata(meta))
+    }
+
+    #[napi]
+    pub async fn create_dir(&self, path: String) -> Result<()> {
+        self.0.create_dir(&path).await.map_err(format_napi_error)
+    }
+
+    #[napi]
+    pub fn create_dir_sync(&self, path: String) -> Result<()> {
+        self.0
+            .blocking()
+            .create_dir(&path)
+            .map_err(format_napi_error)
     }
 
     #[napi]
@@ -186,6 +154,13 @@ impl Operator {
     }
 
     #[napi]
+    pub fn scan_sync(&self, path: String) -> Result<BlockingLister> {
+        Ok(BlockingLister(
+            self.0.blocking().scan(&path).map_err(format_napi_error)?,
+        ))
+    }
+
+    #[napi]
     pub async fn delete(&self, path: String) -> Result<()> {
         self.0.delete(&path).await.map_err(format_napi_error)
     }
@@ -201,9 +176,9 @@ impl Operator {
     }
 
     #[napi]
-    pub async fn writer(&self, path: String) -> Result<Writer> {
-        Ok(Writer(
-            self.0.writer(&path).await.map_err(format_napi_error)?,
+    pub fn list_sync(&self, path: String) -> Result<BlockingLister> {
+        Ok(BlockingLister(
+            self.0.blocking().scan(&path).map_err(format_napi_error)?,
         ))
     }
 }
@@ -220,29 +195,20 @@ impl Entry {
 }
 
 #[napi]
-pub enum EntryMode {
-    /// FILE means the object has data to read.
-    FILE,
-    /// DIR means the object can be listed.
-    DIR,
-    /// Unknown means we don't know what we can do on this object.
-    Unknown,
-}
-
-#[allow(dead_code)]
-#[napi]
 pub struct Metadata(opendal::Metadata);
 
 #[napi]
 impl Metadata {
-    /// Mode of this object.
-    #[napi(getter)]
-    pub fn mode(&self) -> EntryMode {
-        match self.0.mode() {
-            opendal::EntryMode::DIR => EntryMode::DIR,
-            opendal::EntryMode::FILE => EntryMode::FILE,
-            opendal::EntryMode::Unknown => EntryMode::Unknown,
-        }
+    /// Returns true if the <op.stat> object describes a file system directory.
+    #[napi]
+    pub fn is_directory(&self) -> bool {
+        self.0.is_dir()
+    }
+
+    /// Returns true if the <op.stat> object describes a regular file.
+    #[napi]
+    pub fn is_file(&self) -> bool {
+        self.0.is_file()
     }
 
     /// Content-Disposition of this object
@@ -262,13 +228,6 @@ impl Metadata {
     pub fn content_md5(&self) -> Option<String> {
         self.0.content_md5().map(|s| s.to_string())
     }
-
-    // /// Content Range of this object.
-    // /// API undecided.
-    // #[napi(getter)]
-    // pub fn content_range(&self) -> Option<Vec<u32>> {
-    //     todo!()
-    // }
 
     /// Content Type of this object.
     #[napi(getter)]
@@ -314,31 +273,22 @@ impl Lister {
 }
 
 #[napi]
-pub struct Writer(opendal::Writer);
+pub struct BlockingLister(opendal::BlockingLister);
 
+/// Method `next` can be confused for the standard trait method `std::iter::Iterator::next`.
+/// But in JavaScript, it is also customary to use the next method directly to obtain the next element.
+/// Therefore, disable this clippy. It can be removed after a complete implementation of Generator.
+/// FYI: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Generator
 #[napi]
-impl Writer {
-    /// # Safety
-    ///
-    /// > &mut self in async napi methods should be marked as unsafe
-    ///
-    /// napi will make sure the function is safe, and we didn't do unsafe
-    /// thing internally.
+#[allow(clippy::should_implement_trait)]
+impl BlockingLister {
     #[napi]
-    pub async unsafe fn append(&mut self, content: Either<Buffer, String>) -> Result<()> {
-        let c = content.as_ref().to_owned();
-        self.0.append(c).await.map_err(format_napi_error)
-    }
-
-    /// # Safety
-    ///
-    /// > &mut self in async napi methods should be marked as unsafe
-    ///
-    /// napi will make sure the function is safe, and we didn't do unsafe
-    /// thing internally.
-    #[napi]
-    pub async unsafe fn close(&mut self) -> Result<()> {
-        self.0.close().await.map_err(format_napi_error)
+    pub fn next(&mut self) -> Result<Option<Entry>> {
+        match self.0.next() {
+            Some(Ok(entry)) => Ok(Some(Entry(entry))),
+            Some(Err(e)) => Err(format_napi_error(e)),
+            None => Ok(None),
+        }
     }
 }