+++ src/types/operator/blocking_operator.rs
@@ -225,11 +225,14 @@ impl BlockingOperator {
     /// # }
     /// ```
     pub fn metadata(&self, entry: &Entry, flags: impl Into<FlagSet<Metakey>>) -> Result<Metadata> {
-        let meta = entry.metadata();
-        if meta.bit().contains(flags) || meta.bit().contains(Metakey::Complete) {
-            return Ok(meta.clone());
+        // Check if cached metadata saticifies the query.
+        if let Some(meta) = entry.metadata() {
+            if meta.bit().contains(flags) || meta.bit().contains(Metakey::Complete) {
+                return Ok(meta.clone());
+            }
         }
 
+        // Else request from backend..
         let meta = self.stat(entry.path())?;
         Ok(meta)
     }
@@ -258,19 +261,20 @@ impl BlockingOperator {
         }
     }
 
-    /// Create an empty file or dir, like using the following linux commands:
+    /// Create a dir at given path.
     ///
-    /// - `touch path/to/file`
-    /// - `mkdir path/to/dir/`
+    /// # Notes
+    ///
+    /// To indicate that a path is a directory, it is compulsory to include
+    /// a trailing / in the path. Failure to do so may result in
+    /// `NotADirectory` error being returned by OpenDAL.
     ///
     /// # Behavior
     ///
     /// - Create on existing dir will succeed.
-    /// - Create on existing file will overwrite and truncate it.
+    /// - Create dir is always recursive, works like `mkdir -p`
     ///
     /// # Examples
-    ///
-    /// ## Create a dir
     ///
     /// ```no_run
     /// # use std::io::Result;